Extract save dashboard action
