+++ frontend/src/metabase/dashboard/actions/index.ts
@@ -3,6 +3,7 @@ export * from "./core";
 export * from "./data-fetching";
 export * from "./navigation";
 export * from "./parameters";
+export * from "./save";
 export * from "./sharing";
 export * from "./ui";
 export * from "./writeback";
